Rename of XHR - Moving files
