+++ modules/@angular/compiler/src/resource_loader.ts
