+++ modules/@angular/compiler/testing/resource_loader_mock.ts
