+++ modules/@angular/platform-browser-dynamic/src/resource_loader/resource_loader_impl.ts
