+++ modules/@angular/platform-browser-dynamic/test/resource_loader/resource_loader_cache_spec.ts
@@ -12,9 +12,9 @@ import {TestBed, fakeAsync, flushMicrotasks, tick} from '@angular/core/testing';
 import {AsyncTestCompleter, TestComponentBuilder, beforeEach, beforeEachProviders, ddescribe, describe, iit, inject, it, xit} from '@angular/core/testing/testing_internal';
 import {expect} from '@angular/platform-browser/testing/matchers';
 
-import {CachedResourceLoader} from '../../src/xhr/xhr_cache';
+import {CachedResourceLoader} from '../../src/resource_loader/resource_loader_cache';
 
-import {setTemplateCache} from './xhr_cache_setter';
+import {setTemplateCache} from './resource_loader_cache_setter';
 
 export function main() {
   describe('CachedResourceLoader', () => {
