+++ modules/@angular/platform-browser-dynamic/test/resource_loader/resource_loader_impl_spec.ts
@@ -7,7 +7,7 @@
  */
 
 import {AsyncTestCompleter, beforeEach, ddescribe, describe, expect, iit, inject, it, xit} from '@angular/core/testing/testing_internal';
-import {ResourceLoaderImpl} from '../../src/xhr/xhr_impl';
+import {ResourceLoaderImpl} from '../../src/resource_loader/resource_loader_impl';
 
 export function main() {
   describe('ResourceLoaderImpl', () => {
